requirements-dev.txt
black
pylint
pytest
pytest-django
coverage
pytest-cov